Fix Composer PSR-4 deprecation notices
